+++ CHANGELOG.md
@@ -2,6 +2,14 @@
 
 Records breaking changes from major version bumps
 
+## 54.0.0
+
+You should now use `flask routes` in an app, instead of `python application.py
+list_routes` or `python application.py list_external_routes`.
+
+`flask routes` incldues the fully qualified name of the route method, so if you
+want to show only external routes you can use `flask routes | grep '^external'`.
+
 ## 53.0.0
 
 - You must change any uses of `dmutils.api_stubs` to use `dmtestutils.api_model_stubs` instead. The api stubs in this package have been removed.